#2823 - Implement Camunda Impacts of Request a Change to Dependants (#3520)
- Added `appealsStudentDependantsData` to
`load-assessment-consolidated-data` to load the appeal data.
  - Followed the same already in place for other appeals.
- Used "Dependants" script task to calculate the dependants based on
application data or appeal data.
- Added the following variables:
  - `appealsStudentHasDependantsData`
  - `appealsStudentDependantsData`
- Followed the latest appeals process like Additional Transportation.
- Added E2E tests by following the same pattern for other appeal test
cases.
- "Should have calculated student dependants variables assigned with
request a change values when there is a request a change for student
dependants."
- "Should have calculated student dependants variables assigned with
request a change values when there is a request a change for student
dependants with no dependants."
- "Should have calculated student dependants variables assigned with
student data values when there is no request a change for student
dependants."

Screenshot of the Camunda workflow for using appeal or application data

![image](https://github.com/user-attachments/assets/1838a191-b353-4869-976b-820a1a53e6a7)