#3382 - NOA - Displaying $0 award/not eligible  (#3571)
- Created a new e-Cert validation `NoEstimatedAwardAmounts`.
- Change Camunda to allow saving disbursements when non-eligible awards
are present, as shown in the image in the associated PR ticket.
- Added a slot in the `student-page-container` to allow the banner to be
in the expected area.
- Changed how the banner lists its content to allow content other than
plain text to be easily added.
- Added server validations missing while accessing the student endpoint,
we need to ensure the student has access to the application.
- Added server validations while accepting the assessment.
- Rename the API endpoint from `get-application-warnings`. The "HTTP
verb" should not be in the route.
- Change the API endpoint `warnings` to return a DTO.
- Changed the e-Cert eligibility query to return records even if the
disbursement values are absent.
- **_Changed behavior_**: following the same idea, allowed a missing
MSFAA association to be part of the validations. MSFAA is a critical
part of the e-Cert and the idea is to raise the e-Cert issue instead of
the record not being selected. This change generated some issues in E2E
tests that were fixed as part of this PR.

### Centralization logic for e-Cert validation result


https://github.com/bcgov/SIMS/blob/905718e27da8521be9ab955a9803ea94e6292b52/sources/packages/backend/libs/integrations/src/services/disbursement-schedule/e-cert-calculation/e-cert-pre-validation-service-models.ts#L19-L49

### UI Sample


![image](https://github.com/user-attachments/assets/f6d33e8e-456b-405d-bd9a-b6796158ad9e)

Update content with the latest decision, minor title change.


![image](https://github.com/user-attachments/assets/5401f8b5-2c39-4ffe-ac03-bdbd7da54746)