#3223 - Create history/journal tables for auditing purposes (users and students) (#3618)
## Proposed Solution
- Created a generic function `create_history_entry` to be associated
with every table that would need a history.
- Used the tables `sims.users` and `sims.students` as a POC with the
expected structure for any history table.
  - Tables were appended with `_history`.
- New column `history_timestamp` to save the time when the entry was
created.
- New column `history_operation` to save the operation as `INSERT` or
`UPDATE`.
  - Index created on `history_timestamp`.
- `NOT NULL` constraints and relationships were dropped to keep the
table as plain as possible.
- _Migration scripts for the new history table were placed inside the
same folder as the target table. When columns need to be added or
dropped, the idea would be to use the same SQL script file to manage
both tables to keep it simple._

## Sample `sims.user` table where the POC was applied

![image](https://github.com/user-attachments/assets/968e3a94-b7d0-4931-b696-71ec33725a14)

### Pros

- Generic table eliminates the need for the creation of individual
functions to deal with different table structures.
- Creation of new columns would not demand changes in the trigger
function.
- The tables could be individually configured for the triggers, for
instance, a table like `education_programs_offerings` can be configured
to generate history only for the `UPDATE` for performance considerations
during bulk inserts.

### Cons

- When adding new columns we should be mindful about adding the columns
to both tables at the same time to ensure the proper insert order (it
may be in the "pros" list 😄 ).
- If new columns need to be added to the `_history` table we must ensure
they will be introduced at the beginning of the table which would demand
recreation of the `_history` table since there is no mechanism to add a
table in the specific order.

### Migrations rollback evidence


![image](https://github.com/user-attachments/assets/1c0cb9c9-abbb-4385-adbb-3f239323159c)


![image](https://github.com/user-attachments/assets/a0c412ad-d5d2-460c-9120-849e9ac23d71)


![image](https://github.com/user-attachments/assets/431806ca-183a-4d63-83d8-8b0e01e03894)

### Smoke test validations
- Created a new basic BCeID and login as a student with Ministry
approval and checked the history created for `sims.users` and
`sims.students` tables.
- Create a new column in `sims.students` and `sims.students_history` and
the history continues to be generated as expected.

![image](https://github.com/user-attachments/assets/f0334eab-103a-4953-b98d-6e3b82cdf68b)

### Todo

- Implement audit history for remaining tables.