#3223 - Create history/journal tables for auditing purposes (#3632)
This PR is the result a cherry-pick from reverted PRs #3618 and #3627.
These PRs were reverted by #3631 to allow the history tables to be
created following exactly the same column orders from the DB it is being
deployed.
The only real change not reviewed previously is the below. All other
files were kept as they were. The description below is also a
combination from the previous reverted PRs (the rollback evidences are
new).
```sql
CREATE TABLE sims.supporting_users_history AS
SELECT
  CURRENT_TIMESTAMP AS history_timestamp,
  '' :: varchar(50) AS history_operation,
  *
FROM
  sims.supporting_users
WHERE
  false;
```

## Proposed Solution
- Created a generic function `create_history_entry` to be associated
with every table that would need a history.
- Created the history tables for the required tables.
  - Tables were appended with `_history`.
- New column `history_timestamp` to save the time when the entry was
created.
- New column `history_operation` to save the operation as `INSERT` or
`UPDATE`.
  - Index created on `history_timestamp`.
- `NOT NULL` constraints and relationships were dropped to keep the
table as plain as possible.
- _Migration scripts for the new history table were placed inside the
same folder as the target table. When columns need to be added or
dropped, the idea would be to use the same SQL script file to manage
both tables to keep it simple._

## Sample `sims.user` table

![image](https://github.com/user-attachments/assets/968e3a94-b7d0-4931-b696-71ec33725a14)

### Pros

- Generic table eliminates the need for the creation of individual
functions to deal with different table structures.
- Creation of new columns would not demand changes in the trigger
function.
- The tables could be individually configured for the triggers, for
instance, a table like `education_programs_offerings` can be configured
to generate history only for the `UPDATE` for performance considerations
during bulk inserts.

### Cons

- When adding new columns we should be mindful about adding the columns
to both tables at the same time to ensure the proper insert order (it
may be in the "pros" list 😄 ).
- If new columns need to be added to the `_history` table we must ensure
they will be introduced at the beginning of the table which would demand
recreation of the `_history` table since there is no mechanism to add a
table in the specific order.

#### Minor Audit-Related Fix

- Supporting users were not saving the audit user (system user) during
the record creation.
- While providing supporting user data, the user was not saved when the
authenticated user was new, which means the user token ID was never
populated leaving the `modifier` blank.

### Migrations rollback evidence


![image](https://github.com/user-attachments/assets/469b86ef-497a-41e6-b898-c5a1229470b4)


![image](https://github.com/user-attachments/assets/5a5c6646-caa9-4e91-adcf-99c4b9cfe6d8)


![image](https://github.com/user-attachments/assets/9d1349cb-96f1-4c6e-81ef-303134dc73df)


![image](https://github.com/user-attachments/assets/eb80b864-08fc-4f06-85cb-e78b4c551877)


![image](https://github.com/user-attachments/assets/e733f4bc-1798-41c3-a928-36fb2f17238d)


![image](https://github.com/user-attachments/assets/edf6c751-2689-4711-911c-2ef0dd5c9975)


![image](https://github.com/user-attachments/assets/e2d81691-bef3-4624-86d4-956aac9c3a7d)


![image](https://github.com/user-attachments/assets/098518b6-0d8c-4c88-bf64-8e0b757a78e8)