#3193 - New OCP Routes (#3649)
## Main changes
- Added `certificate`, `key`, and `caCertificate` to routes.
- Added the proposed annotations to the same routes. Information about
the route annotations and its intentions can be found in
[route-specific-annotations_route-configuration](https://docs.openshift.com/container-platform/4.15/networking/routes/route-configuration.html#nw-route-specific-annotations_route-configuration).
- Renamed the routes (now sims.studentaidbc.ca) to allow both routes to
exist at the same time for now.
- Only the certificate contents are provided as parameters to allow the
correct line-breaking.

### Additional Steps
- Change Keycloak properties for: `aest`, `institution`, `student`, and
`supporting-users`.
  -  Valid Redirect URIs: `https://dev.sims.studentaidbc.ca/*`
  - Web Origins: `https://dev.sims.studentaidbc.ca`
- Created a new GitHub variable `DOMAIN_PREFIX` (created as a regular
var because it is not sensitive).
- Create the new variables `TLS_CERTIFICATE`, `TLS_KEY`,
`TLS_CA_CERTIFICATE`. These were not added to the sample config because
there is no reason for them to exist in the local environment.

### Non-directly related change
- `FORMIO_ROOT_EMAIL` and `FORMIO_ROOT_PASSWORD` are no longer
regenerated at every form.io server deployment to avoid credentials
reconfigurations along the applications. The credentials should be
changed due to a password rotation process and not during the server
update.

### TODO
- Ensure release instructions for the `Additional Steps` above.