#1815 - Ignore restrictions for specific disbursement(s) within an assessment (#3679)
- Created the `restrictionBypassesResolutionStep` to resolve active
bypasses when needed as the last step in the e-Cert generation process
for part-time and full-time.
- Changed the `getEligibleDisbursements` query adding the information
from the active bypasses and the restriction bypassed.
- Method `getRestrictionByActionType` was changed to use the effective
restrictions (active restrictions not bypassed) instead of the active
restrictions.
- The BCLM step is ignored at the beginning of its execution. If the
BCLM restriction is in place the method would not be executed (logic
already in place). The condition based on the bypass was added either
way to ensure that whatever happens to the BCLM logic the bypass will
guarantee that the store step will be ignored.
- A log was created to list all the active restrictions known during the
e-Cert calculation. The log records the restriction code and the student
restriction ID for easier troubleshooting.
- Created factories for the new table `ApplicationRestrictionBypass`.

### E2E Tests

#### Full-time
- Should generate BC awards amounts with no restriction deductions and
resolve the restriction bypass when a student has an active 'BCLM'
restriction and it is bypassed with behavior 'Next disbursement only'.
- Should have the e-Cert generated for a full-time application and the
bypass active when a student has an active 'Stop full time disbursement'
restriction and it is bypassed with behavior 'All disbursements'.
- Should prevent an e-Cert generation and keep the bypass active when
multiple 'Stop full time disbursement' restrictions exist and only one
is bypassed and it is bypassed with behavior 'Next disbursement only'.

#### Part-time
- Should have the e-Cert generated for a part-time application and the
bypass resolved when a student has an active 'Stop part time
disbursement' restriction and it is bypassed with behavior 'Next
disbursement only'.
- Should prevent an e-Cert generation and keep the bypass active when
multiple 'Stop part time disbursement' restrictions exist and only one
is bypassed and it is bypassed with behavior 'Next disbursement only'.

### Log sample without the bypass


![image](https://github.com/user-attachments/assets/cc75c575-df20-49b3-aefc-ebaac508b9f3)

### Log sample with a bypass in place.


![image](https://github.com/user-attachments/assets/be96088b-458e-41d2-abf1-f56d5d610d1b)

### Noted generated when a bypass is resolved by the system

AC was adjusted later to include the application number.

![image](https://github.com/user-attachments/assets/49f3180e-ae52-4f08-b169-9fe5ac93628d)