#3131 - Fix content length to be based on byte lenth (#3695)
# FIX CONTENT LENTH HEADER ON REPORTS CONTROLLER

## Problem statement
Currently, in reports controller we set the `Content-Length` header
value as length of the csv string.
Due to this, we have an issue when the string involves special
characters as the size of these characters are more than one byte.

e.g. When a string of length 100 characters has one special character
like `â` in it, then string length is `100` but the byte length is
`101`. Hence when this is set in API response it eliminates any
characters which is more than set content length.

Like this


![image](https://github.com/user-attachments/assets/565166cc-8b51-4345-8da5-9da87b07e02e)


## Solution
Set `content-length` header with byte length.