#2509 - Update formio server and js - Side Effects (#3747)
Between version 3.55 and 4.00 the server validation changed in a way
that affected the calculated inputs.

First, for calculated inputs participating in validations, the value
should be set as below to avoid being manipulated by a malicious user.
For instance for a hidden input with a fixed value.
- "defaultValue": null (not required)
- "calculateValue": "value = \"100000000\";"
- "calculateServer": true
- "persistent": true | false, (it must be set to false if there is no
point in saving it).

![image](https://github.com/user-attachments/assets/e942b95c-5340-425f-b951-d5c5d87af651)

In version 4.0.0, the order in which the calculated input is added to
the screen started to matter in some scenarios. Even declaring
calculated inputs as the above, if the calculated input is declared
after the component that consumes its value, it would allow the user to
manipulate the hidden input for calculations (the stored value would be
right either way).

### PR intended fixes
- Adjusted the Student Application and Supporting Users forms to follow
the recommended approach above and also moved some hidden fields before
the components using their calculated value. This change affects mostly
`maxIncome` and `maxMoneyValue` validations.

_Example of some hidden fields that were just moved in the form._

![image](https://github.com/user-attachments/assets/078b2f63-a18a-42b2-8690-63df090b27f0)

- Some fields were not updating their content when the content had
expressions. The solution was to add `"refreshOnChange": true`. The
issue was noticed in appeals and added to the application as a
precaution.
- Dependents table in Student Application and Supporting Users Parents
were having an issue where the submission was not allowed because the
entire grid was `required: true`. Removing the `required: true` from the
grid allowed the form to still be validated and the form submitted.

_Please note that logics like below were not changed and are marked just
because the `isNew`_.

![image](https://github.com/user-attachments/assets/32d1d928-d146-46bb-b1dc-ef7fcd634648)