#2568 -  Minimum validation for the date entered on scholastic standing creation (#3766)
- [x] Added validation in the formio "Report Scholastic Standing Change"
on the date parameters for Scholastic Standings when they are created
- [x] API validation is also changed with respect to the formio
validation, so the logic which fetches the study start and end date has
been included and the service is moved with respect to it, without
making additional calls.
- [x] The archived check for the list of applications shown in the
'Available to Report ' tab is already implemented.
- [x] E2E test cases where expanded and included for the scholastic
standing related to date like School Transfer, Student Withdrawal and
Student completed early with the dates between the study start , end
period and current date.
- [x] Added E2E test cases for the list that shows active applications
in Report a Change 'Available to Report' tab.

---------

Co-authored-by: guru-aot <gurumoorthy.mohan@gov.bc.ca>