#3260 - CAS Integration 3A - Create new Supplier and Site - Refactor CAS authentication (#3822)
As agreed in the previous PR, the CAS authentication is now encapsulated
inside the CAS API service and the auth token no longer needs to be
passed to every method.
The current CAS token expiration is set to 1 hour and the token will be
considered "about to expire" one minute before it.