#3260 - CAS Integration 3A - Create new Supplier and Site - Retry on Error and Scheduler Time (#3830)
- Changed the to allow throwing an error when some error is present in
the process summary error attending the below AC.
```
If some error is present throw in the scheduler to allow it to be retried.
```
- Added the below E2E to test the above change.
- Should throw an error for the first student and process the second one
when the CAS API call failed for the first student but worked for the
second one.
- Adjusted the scheduler to be executed during weekdays only attending
the below AC (`0 20 * * 1-5`).
```
Schedule these to go once a day at noon during weekdays for now.
```

### Rollback

![image](https://github.com/user-attachments/assets/59a242ed-72ec-465b-a52f-8848144568f0)

### Fix
- `COUNTRY_CANADA` was defined as "canada" when it is defined as
"Canada" is all form.io definitions leading to an incorrect logic.
- SFTP issue while providing the file content as a string instead of a
Buffer as identified by @dheepak-aot.