#3774 - SFAS MSFAA Implementation (#3817)
- [x] Refactored associateMSFAANumber in
disbursement.schedule.service.ts as the Cognitive Complexity became 18
so changed for early return during creation or assigning of MSFAA.
- [x] During Associate MSFAA, if valid MSFAA number is present for a
student, then use it.
- [x] Else check for previously signed application disbursement already
sent with valid end date and MSFAA number is present, use that.
- [x] Else check if SFAS individual MSFAA number is present and checked
if the application end date is within validate date, then use the MSFAA
and create in SIMS and activate it.
- [x] Else If MSFAA is not found in SFAS application, then check if
MSFAA number is present in pending state, waiting for signing.
- [x] Else create MSFAA number and activate and assign it to the
application.
- [x] E2E Test cases created for the above scenarios.
- [x] Removed a not useful check in the getCurrentValidMSFAANumber
method, which has referenceApplication as not null.
- [x] The Sonarcloud issue is related to Cognitive Complexity in test
helper class, so ignored.

---------

Co-authored-by: guru-aot <gurumoorthy.mohan@gov.bc.ca>