#3889 - Increase Logging for SFTP and file handling (#3897)
## Main PR Goal

- Increase the log level at the `SFTPIntegrationBase` without disrupting
the consumer's methods and try to limit the amount of refactoring.
- Deploy to PROD in the next possible release to allow troubleshooting
of current intermittent issues while archiving files.

### Changes

- Adapted the `LoggerService.error` to receive a friendly error and log
the error itself, keeping it compatible with the existing methods.
- Added log messages to all base SFTP operations in the
`SFTPIntegrationBase`. The errors are reshown keeping the current
behavior for current method consumers.
- Adjusted the API `mais.ts` that was using the log context.
- Adjusted the global error handler to generate only one log entry
instead of three.
- Even though the `SFTPIntegrationBase` method is now generating more
logs, the archive operation error handling was adjusted to ensure the
exception details are logged.
- Small refactor to stop passing the `SFTP_ARCHIVE_DIRECTORY` for every
archive method.