#2741 - Ministry Filter: Search Programs and Offerings [e2e tests] (#3883)
### As a part of this PR, the following were completed:

- **Wrote the following e2e tests:**

**EducationProgramAESTController(e2e)-getProgramsSummary**
    
√ Should not retrieve any education programs when the search for the
programs is made and no program statuses are searched for including no
search for the inactive programs. (655 ms)
√ Should retrieve all the education programs when the search for the
programs is made and all program statuses are searched for including the
inactive programs. (385 ms)
√ Should retrieve all the inactive education programs when the search
for the programs is made and none of the program statuses are selected
except for the inactive programs. (359 ms)
√ Should retrieve all the pending, declined and inactive education
programs sorted by ascending program status when the search for the
programs is made and pending, declined program statuses are selected
along with the inactive programs. (380 ms)
√ Should retrieve all the approved and declined education programs when
the search for the programs is made and only the approved and declined
program statuses are selected. (349 ms)
√ Should retrieve all the approved and pending education programs
matching the program name search string when the search for the programs
is made and the approved and pending program statuses are selected along
with the program name search filter. (370 ms)
√ Should retrieve all the declined and inactive education programs
sorted by ascending program status matching the program name search
string and the location name search string when the search for the
programs is made and the declined program status is selected along with
the inactive programs selected and the program name and location name
search filter. (623 ms)


------------------------------------------------------------------------------------------------------

### Screenshot of the executed tests:

<img width="1452" alt="image"
src="https://github.com/user-attachments/assets/ae7df65f-03d1-46b0-996f-ae4b540af9c8">


------------------------------------------------------------------------------------------------------

- **Fixed the issue where the checkbox doesn't show for the `v-select`
component when the list item is unchecked.**


**Issue:**

<img width="1920" alt="image"
src="https://github.com/user-attachments/assets/95403598-bd41-446e-b216-66e41900387c">


------------------------------------------------------------------------------------------------------

**Fix:**

<img width="1918" alt="image"
src="https://github.com/user-attachments/assets/4758be7b-c788-48cd-9f43-dd376ad9dd88">