#3789 - PIR: Program Name and Program Description Bug (#3944)
**As a part of this PR, the bug for the PIR where the Program Name and
the Program Description are not showing up once the application is
submitted is fixed.**

**Fix:** The bug was happening because both the fields (Name and
Description) were having a conditional set as below:

```
"conditional": {
  "show": true,
  "when": "programnotListed",
  "eq": "true"
},
```

Removed the above piece of code for both the program name and program
description for all the program years.

The problem is that `programnotListed` never has the value `true`. It is
a `Select Boxes Component` rather than the `Checkbox Component` which
gets the value `programnotListed`. We do not need to rely on this
component because both the components `Program Name` and `Program
Description` are wrapped inside a container component that has a
conditional setup on it.

Tested the above for the below scenarios:

**1.  Program is not listed**

<img width="1025" alt="image"
src="https://github.com/user-attachments/assets/4e14f404-af5e-4306-b970-b7859cbce5a6">


-------------------------------------------------------------------------------------------------------------

**2.  Study is not listed**

<img width="944" alt="image"
src="https://github.com/user-attachments/assets/926615d5-6e12-4acb-888e-152be523c166">