#3667 - Bull Scheduler Increments the Next Scheduled Job When Manually Promoted (#3957)
Updated Bull Board-related packages to take advantage of the feature
present in the newer version that allows a new job to be added (same as
duplicate), allowing the schedulers to be executed without affecting
their current time to be executed again.

_Notes:_
1. during the local tests, when a delayed job was deleted and the
`queue-consumers` restarted, the delayed job was restored.
2. this PR is intended to be part of the upcoming release. The research
for the ticket will continue.

## Using the add/duplicate option

While creating the new job the properties can be edited, for instance,
the corn expression can be edited to create a new time to execute the
scheduler.

_Please note that `timestamp` and `prevMillis` should be removed. These
properties are generated once the job is created.
As mentioned in the source code, `timestamp` is the "Timestamp when the
job was created." and `prevMillis` is a "Internal property used by
repeatable jobs."_


![image](https://github.com/user-attachments/assets/05245952-5bda-48db-a3b6-61e3a2317669)

```json
{
  "repeat": {
    "count": 1,
    "key": "__default__::::0 7 * * *",
    "cron": "0 7 * * *"
  },
  "jobId": "repeat:2c2720c5e8b4e9ce99993becec27a0ff:1731999600000",
  "delay": 42485905,
  "timestamp": 1731957114095,
  "prevMillis": 1731999600000,
  "attempts": 3,
  "backoff": {
    "type": "fixed",
    "delay": 180000
  }
}
```

## Refactor

- Refactored the code to move the Bull Board configuration to the Nestjs
modules instead of doing it on the main.ts.
The refactored code is equivalent to the one previously on the
`main.ts`.
- Updated icons and labels to make the dashboard look more like part of
the SIMS. The way it was done was by targeting less effort. In case it
causes some noise during the PR review the code will be removed.