#3789 - Update the form IO conditions based on form data (#3964)
## Updated the condition to render the program name and program
description

- [x] Updated the condition to render the program name and program
description fields only when a program is not selected in the student
application(One of the PIR situation).
- [x] The input which says `Your program is not listed` is a `select
box` and not check box. So we can use only custom condition to evaluate
its output at it will be in an object form like this.
```json
 "myProgramNotListed": {
      "programnotListed": false
    }
```


![image](https://github.com/user-attachments/assets/f3d55d2a-4455-4bed-883d-b5d78e0c1cc8)

**Formio condition updated to :**

![image](https://github.com/user-attachments/assets/305385c3-30ac-4434-b5d2-c5d9a9c74227)


## Updated the read only selected program and offering display

- [x] Updated the read only selected program and selected offering
display only when the respective value is present.

Currently it is displaying as blank in PIR situations.


![image](https://github.com/user-attachments/assets/00e42823-5e4c-4f67-a0aa-5d5864e7a478)


![image](https://github.com/user-attachments/assets/d5fa0abc-c7ab-4e00-b5a1-96dc189299b2)