#3789 - Fix program description on Institution PIR Page (#3967)
## Issue description and fix

Issue: When a program is selected in a student application but study
period is not selected and submitted for PIR review, the Institution PIR
page was displaying the `program name` in the `program description`
section.

- [x] Fix: The formIO was neither receiving nor displaying a program
description that was coming from the original program. Hence the API was
enhanced and FormIO was updated to receive the program description
variable.


![image](https://github.com/user-attachments/assets/a9e9843a-c18e-4de8-b1a7-3b3373380cde)


![image](https://github.com/user-attachments/assets/5f1944df-5340-4ef7-b471-ab89cf3dd625)