#3667 - Bull Scheduler increments the next scheduled job when manually promoted (Ensure Next Delayed Job) (#3978)
- Ensures a scheduled job will have a delayed job created with the next
expected scheduled time based on the configured cron expression.
- DB row used as a lock to ensure the initialization code will not
recreate delayed jobs or create them in an unwanted way if two PODs
start at the same time. To be clear, the lock approach is the same used
in other areas of the application and locks a single row in the table
queue-consumers, please not, the table will not be locked. Please see
below the sample query executed.
```sql
START TRANSACTION
select
	"QueueConfiguration"."id" AS "QueueConfiguration_id"
from
	"sims"."queue_configurations" "QueueConfiguration"
where
	(("QueueConfiguration"."queue_name" = 'archive-applications')) LIMIT 1 FOR UPDATE 
COMMIT
```
- While the queue is paused, no other delayed jobs will be created.