#3881 - Validate user account for all routes (#3985)
# Validate user account for all routes

## New global guard and decorator

- [x] New guard `RequiresUserAccountGuard` has been introduced globally
to ensure that routes are authorized with the user token which belongs
to valid SIMS user. There are exceptional routes like public routes and
routes used that setup the user itself are skipped from this validation.
- [x] New decorator `@RequiresUserAccount()` is introduced to get the
metadata context for the guard.

## Student page container

- [x] Student page container updated to NOT render restriction and SIN
banners for pages which does not require a valid student account.

## E2E Tests

- [x] The existing method to mock the student info from token
`mockUserLoginInfo()` does not have a way to restore the mock, if the
mock needs to be restored for other tests in same suite.
Hence refactored the code to use `jest.spyOn()` to mock the userService
method implementation and also created a reset mock method to reset the
mock as required in the test suite.

Here is an example.
**Mock applied**


![image](https://github.com/user-attachments/assets/9f0277d8-27c9-4c1a-98d0-8fd5fc5583a8)

**Mock Reset**

![image](https://github.com/user-attachments/assets/0e9115b3-d456-43eb-aace-d282d68d5297)

- [x] Created new Auth E2E tests


![image](https://github.com/user-attachments/assets/b3e48542-c314-4ee7-81e3-6807a6ea32f6)


## Volar extension

- [x] Updated the workspace file with deprecated vue extension by
replacing with recommended extension.

![image](https://github.com/user-attachments/assets/1ee3de6f-b315-4d87-acd2-32f58dce0891)