#3666 - Queue Monitoring - Default Queue Error Handling (#4020)
This PR is a proposal to unify the queue behavior. It contains changes
for start and cancel assessments regular queues and the idea is to
expand it to all schedulers.

The below image shows these points to be shared and forced for all the
queues, using the CAS scheduler as an example.
1 - Return type;
2 - Provide a `processSummary`;
3 - try/catch and default start message.
4 - Error check on `processSummary` to force the jobs to fail.
5 - Error handling on `catch`.
6 - Final log activities during `finally`.

![image](https://github.com/user-attachments/assets/4b31491b-f859-4ef6-a126-67ad60bf7bb0)
_Note:_ The above will also ensure the E2E tests follow a closer
approach.

Sample log for a queue using the new `BaseQueue<T>` class.


![image](https://github.com/user-attachments/assets/f42193dc-5e57-4f8a-94a8-c4a78df39084)

### Side effects
The idea is to have the BaseScheduler extend the BaseQueue which will
force a change in every single scheduler (which would be the idea). To
avoid the need to change every single scheduler and adapt every E2E, the
methods required by BaseQueue can be implemented temporarily as below.
The idea is to start the scheduler refactor still during this ticket.

```ts
/**
 * To be implemented.
 */
processQueue(_job: Job<void>): Promise<string | string[]> {
  throw new Error("Method not implemented.");
}

/**
 * To be implemented.
 */
async process(
  _job: Job<void>,
  _processSummary: ProcessSummary,
): Promise<string | string[]> {
  throw new Error("Method not implemented.");
}
```

### Error Fix
Fixed an issue with the logger service where `null` message was logged
right after the regular message.