#4087 - Search Student Bug (#4089)
The create student action happening when Ministry user trying to approve
a student. The SIN validation table is not inserted. This causes the
search to fail as it uses the below query to fetch.


![image](https://github.com/user-attachments/assets/d0138e14-e475-478e-9435-cbf944db3ebf)

This is a bug created during the cas supplier pending verification logic
centralization that happened in the #3258 ticket.

Co-authored-by: guru-aot <gurumoorthy.mohan@gov.bc.ca>