#4132 - Update Form IO (#4148)
- Updated form IO server version from 4.2.4 to 4.3.0 
- Upgraded `"formiojs": "^4.21.4"` to `"@formio/js": "^5.0.0"` for Web
App
- Fixed user interface issues due to the upgrade in Web App
- Added CSS to remove extra text on file component that were hidden in
previous versions.
- Added CSS to show the "remove icon" and "calendar icon" which were
hidden in the current version.
- Changed HTML components to **Text Field Components** for displaying
the awards on the Notice of Assessment page.
- Changed `fileMaxSize` from "4 MB" to "15MB" for the file component
`pdDependentUpload` for all program years.
- Reverted `clearOnHide` for the file component
`dependantCustodyFileUpload` for sfaa2025-2026.

Screenshot of the file component UI before and after the fix

![image](https://github.com/user-attachments/assets/ec03be17-f7fc-437c-ab82-7a92c5896e9a)

Screenshot of the calendar icon before and after the fix

![image](https://github.com/user-attachments/assets/bd1f78f8-8ed0-4500-bb0e-1ef278c8eed6)

Screenshot of the NOA page due to Form.io upgrade without the fix

![image](https://github.com/user-attachments/assets/7e81ab9d-939b-4063-b497-82ec21879acc)

Screenshot of the NOA page using Text Field components to fix the
rendering issues

![image](https://github.com/user-attachments/assets/8497734b-168d-4058-81c1-68d71aec98d0)

---------

Co-authored-by: Lewis Chen <lewis.chen@aot-technologies.com>