#3666 - Queue Monitoring - Sysdig Team Update (#4233)
This ticket is to add missing team members to the sysdig as agreed
during the development of the ticket #3666.
_Please note the ticket is not associated with the PR because it is
closed and does not seem worth opening or creating a ticket only for
this task._
The current way to update and apply the changes was manual and demanded
manual editing of the file to be applied to each license plate. The
intention of this PR is to make it simple.

- Converted sysdig team to a template to allow adding a parameter.
- Create a make command to execute the template in the current two
license plates.

Technical documentation about Sysdig team setup:
https://developer.gov.bc.ca/docs/default/component/platform-developer-docs/docs/app-monitoring/sysdig-monitor-setup-team/?utm_source=digital&utm_medium=web&utm_campaign=sysdig-monitor